+++ semana20/jest2/respostas.md
@@ -1,0 +1,2 @@
+### Exercício 3
+    c) A segunda função eu posso pular o verificador se eu quiser
