+++ semana20/jest2/src/model/fighter.ts
@@ -1,0 +1,6 @@
+export interface Fighter {
+    name: string,
+    life: number,
+    strength: number,
+    defense: number
+}
