+++ semana20/jest2/src/perfomeAttack.ts
@@ -1,0 +1,29 @@
+import { Fighter } from './model/fighter'
+import { validateFighter } from './validadeCharacter'
+
+export const perfomAttack1 = (attacker: Fighter, defender: Fighter) : void => {
+
+    if (!validateFighter(attacker) || !validateFighter(defender)) {
+
+        throw new Error("Lutador Inválido")
+    }
+
+    if (attacker.strength > defender.defense) {
+
+        defender.life -= attacker.strength - defender.defense
+    }
+}
+
+export const perfomAttack2 = 
+(attacker: Fighter, defender: Fighter, validator: (fighter: Fighter) => boolean) : void => {
+
+    if (!validator(attacker) || !validator(defender)) {
+
+        throw new Error("Lutador Inválido")
+    }
+
+    if (attacker.strength > defender.defense) {
+
+        defender.life -= attacker.strength - defender.defense
+    }
+}
